test: migrate find_free_port to pop_common
